+++ Prototype/ObjectManager/ObjectManager.h
@@ -53,12 +53,12 @@
 // internal handler
 - (void) handlerForCreate:(id)result;
 // request create object
-+ (uint32_t) createObjectWithHandler:(SEL)handler andTarget:(id)target;
++ (NSInteger) createObjectWithHandler:(SEL)handler andTarget:(id)target;
 
 // update method
 - (void) handlerForUpate:(id)result;
 // request update object
-+ (uint32_t) updateObjectWithhandler:(SEL)handler andTarget:(id)target;
++ (NSInteger) updateObjectWithhandler:(SEL)handler andTarget:(id)target;
 
 
 // functions that must be overwrite by sub class if you want to use that method
