+++ Prototype/View/ShareNewEvent/CreateFoodPage.h
@@ -12,5 +12,5 @@
 @interface CreateFoodPage : UITableViewController 
 - (void) imageUploadCompleted:(id)result;
 - (void) needScrollToBegin;
-- (void) resetImageWithUploadFileID:(uint32_t)fileID;
+- (void) resetImageWithUploadFileID:(NSInteger)fileID;
 @end
